pyproject.toml
[project]
name = "multi-agent-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-core>=1.3.2",
    "langchain-nvidia-ai-endpoints>=1.2.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
]
